added auto refresh to lofile list in webapp
